requirements.txt
Flask
google-cloud-firestore
Firestore
